# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification Service API",
        "description": "Микросервис для массовой рассылки SMS/Email уведомлений с приоритетами",
        "contact": {
            "email": "support@notificationservice.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер"
        },
        {
            "url": "https://api.notificationservice.com",
            "description": "Production сервер"
        }
    ],
    "paths": {
        "/api/v1/notifications/send": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Массовая рассылка уведомлений",
                "description": "Отправляет SMS или Email сообщения нескольким получателям",
                "operationId": "sendMassNotifications",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendMassNotificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Уведомления успешно поставлены в очередь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SendMassNotificationResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Запрос уже был обработан (дедубликация)"
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/api/v1/subscribers/{subscriberId}/history": {
            "get": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "История уведомлений подписчика",
                "description": "Возвращает все уведомления для указанного подписчика",
                "operationId": "getSubscriberHistory",
                "parameters": [
                    {
                        "name": "subscriberId",
                        "in": "path",
                        "description": "Идентификатор подписчика (телефон или email)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 500
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subscriber_id": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notification"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/health": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Проверка состояния сервиса",
                "description": "Возвращает статус микросервиса",
                "operationId": "healthCheck",
                "responses": {
                    "200": {
                        "description": "Сервис работает",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "service": {
                                            "type": "string",
                                            "example": "Notification Service"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Notification": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "idempotency_key": {
                        "type": "string",
                        "example": "unique_123"
                    },
                    "subscriber_id": {
                        "type": "string",
                        "example": "+79001234567"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email"
                        ],
                        "example": "sms"
                    },
                    "message": {
                        "type": "string",
                        "example": "Test message"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "queued",
                            "sent",
                            "delivered",
                            "dropped"
                        ],
                        "example": "queued"
                    },
                    "retry_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "SendMassNotificationRequest": {
                "required": [
                    "channel",
                    "message",
                    "recipients",
                    "idempotency_key"
                ],
                "properties": {
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email"
                        ],
                        "example": "sms"
                    },
                    "message": {
                        "type": "string",
                        "maxLength": 5000,
                        "example": "Your verification code: 123456"
                    },
                    "recipients": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "maxItems": 1000,
                        "minItems": 1,
                        "example": [
                            "+79001234567",
                            "+79007654321"
                        ]
                    },
                    "priority": {
                        "type": "integer",
                        "maximum": 10,
                        "minimum": 0,
                        "example": 5
                    },
                    "idempotency_key": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "unique_request_id_123"
                    }
                },
                "type": "object"
            },
            "SendMassNotificationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Notifications queued successfully"
                    },
                    "data": {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 2
                            },
                            "dispatched": {
                                "type": "integer",
                                "example": 2
                            },
                            "channel": {
                                "type": "string",
                                "example": "sms"
                            },
                            "priority": {
                                "type": "integer",
                                "example": 5
                            },
                            "idempotency_key": {
                                "type": "string",
                                "example": "unique_request_id_123"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object"
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Notifications",
            "description": "Endpoints для отправки уведомлений"
        },
        {
            "name": "Subscribers",
            "description": "Endpoints для истории подписчиков"
        },
        {
            "name": "System",
            "description": "Системные endpoints"
        }
    ]
}